# ssl-application/target-migrated-backup/generated-sources/openapi/src/main/java/com/example/ssl/openapi/generated/api/TasksApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.14.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.ssl.openapi.generated.api;

import com.example.ssl.openapi.generated.model.ErrorResponse;
import com.example.ssl.openapi.generated.model.TaskItemPayload;
import com.example.ssl.openapi.generated.model.TaskItemUpsertPayload;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.lang.Nullable;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-20T10:43:47.615777466+01:00[Europe/Madrid]", comments = "Generator version: 7.14.0")
@Validated
@Tag(name = "Tasks", description = "the Tasks API")
public interface TasksApi {

    /**
     * POST /api/tasks : Crear task
     *
     * @param taskItemUpsertPayload  (required)
     * @return Task creada (status code 201)
     *         or Petición inválida (status code 400)
     */
    @Operation(
        operationId = "createTask",
        summary = "Crear task",
        tags = { "Tasks" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Task creada", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TaskItemPayload.class))
            }),
            @ApiResponse(responseCode = "400", description = "Petición inválida", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/tasks",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<TaskItemPayload> createTask(
        @Parameter(name = "TaskItemUpsertPayload", description = "", required = true) @Valid @RequestBody TaskItemUpsertPayload taskItemUpsertPayload
    );


    /**
     * DELETE /api/tasks/{id} : Eliminar task
     *
     * @param id  (required)
     * @return Eliminada (status code 204)
     *         or No encontrada (status code 404)
     */
    @Operation(
        operationId = "deleteTask",
        summary = "Eliminar task",
        tags = { "Tasks" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Eliminada"),
            @ApiResponse(responseCode = "404", description = "No encontrada", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/tasks/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<Void> deleteTask(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    );


    /**
     * GET /api/tasks/{id} : Obtener task por id
     *
     * @param id  (required)
     * @return Task encontrada (status code 200)
     *         or No encontrada (status code 404)
     */
    @Operation(
        operationId = "getTaskById",
        summary = "Obtener task por id",
        tags = { "Tasks" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Task encontrada", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TaskItemPayload.class))
            }),
            @ApiResponse(responseCode = "404", description = "No encontrada", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/tasks/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<TaskItemPayload> getTaskById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    );


    /**
     * GET /api/tasks : Listar tasks
     *
     * @return Lista de tareas (status code 200)
     */
    @Operation(
        operationId = "listTasks",
        summary = "Listar tasks",
        tags = { "Tasks" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Lista de tareas", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = TaskItemPayload.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/tasks",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<TaskItemPayload>> listTasks(
        
    );


    /**
     * PUT /api/tasks/{id} : Actualizar task
     *
     * @param id  (required)
     * @param taskItemUpsertPayload  (required)
     * @return Task actualizada (status code 200)
     *         or Petición inválida (status code 400)
     *         or No encontrada (status code 404)
     */
    @Operation(
        operationId = "updateTask",
        summary = "Actualizar task",
        tags = { "Tasks" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Task actualizada", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TaskItemPayload.class))
            }),
            @ApiResponse(responseCode = "400", description = "Petición inválida", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "No encontrada", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/tasks/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<TaskItemPayload> updateTask(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id,
        @Parameter(name = "TaskItemUpsertPayload", description = "", required = true) @Valid @RequestBody TaskItemUpsertPayload taskItemUpsertPayload
    );

}